Завершает ДЗ
