+++ js/thumbs.js
@@ -1,4 +1,5 @@
 import {createObjects} from './create-object.js';
+import './form.js';
 
 const usersPhotos = document.querySelector('.pictures');
 
@@ -8,7 +9,7 @@ const picturesUsersTemplate = document.querySelector('#picture')
 
 const similarPhotos = createObjects();
 
-const similarPhotosAbove = () => {
+const createSimilarPhotosAbove = () => {
   similarPhotos.forEach((user) => {
     const pictureElement = picturesUsersTemplate.cloneNode(true);
     pictureElement.querySelector('.picture__img').src = user.url;
@@ -19,27 +20,4 @@ const similarPhotosAbove = () => {
   });
 };
 
-const uploadFiles = document.getElementById('upload-file');
-const formEditImg = document.querySelector('.img-upload__overlay');
-const body = document.querySelector('body');
-const uploadFilesClose = document.getElementById('upload-cancel');
-
-uploadFiles.addEventListener('change', () => {
-  formEditImg.classList.remove('hidden');
-  body.classList.add('modal-open');
-});
-
-uploadFilesClose.addEventListener('click', () => {
-  formEditImg.classList.add('hidden');
-  body.classList.remove('modal-open');
-});
-
-document.addEventListener('keydown', (evt) => {
-  if (evt.key === 'Escape') {
-    evt.preventDefault();
-    formEditImg.classList.add('hidden');
-    body.classList.remove('modal-open');
-  }
-});
-
-export {similarPhotosAbove};
+export {createSimilarPhotosAbove};
